@@ -1,14 +1,24 @@
-import CurrencyBox from './components/CurrencyBox/CurrencyBox';
-import Header from './components/Header/Header';
-import Container from './components/Container/Container';
+import logo from './logo.svg';
+import './App.css';
 
-const App = () => {
-
+function App() {
   return (
-    <Container>
-      <Header />
-      <CurrencyBox />
-    </Container>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
   );
 }
 
